Update Shyp/waterline to v1.2.0
We've made a bunch of changes, so let's test and see how they're doing. It
might make sense to just point this at master, and we might want to hook this
up to the `waterline` project but I want to see how it goes as is/gather more
data points before making a decision.

Removing 'alter' was a huge pain since pretty much every test relied on
implicit database creation via the 'alter' mechanism. To replace this I tailed
the Postgres database logs to catch the CREATE TABLE calls, then wrote the SQL
calls to files. These are loaded when the test fixtures run, instead of
'alter'.

I had to delete a bunch of tests which don't pass any more, notably the
`findOrCreateEach` tests and the `findOrCreate` tests. Furthermore the
many-to-many tests and the hasManyThrough tests were deemed to be too painful
to implement, and we don't use them, so I deleted them.

Also un-skips a test since part of the 1.2.0 change should fix the test to do
the right thing.