Replace assert(!err) with assert.ifError(err)
If an error is present, `assert(!err)` reports that false was expected to be
true, which means you have to re-run the test and print the error to actually
figure out what the error was. `assert.ifError(err)` fails if the error is
defined and will print the error if one is present, which means you don't have
to re-run the tests to figure out what went wrong.