Make table names match identity names
The recent changes to error on UPDATE if it matches a primary key require
the table to be named the same as the `identity` - you can't use a different
`tableName`. This is fine, we don't want to support this feature.

Essentially the change here is to make the `tableName` match the identity, and
then update the table names in the `.sql` files to match the `identity`. This
makes waterline-sequel not error.

Removes some tests which now fail - they involve creating nested records by
issuing an UPDATE against the parent. Again, this is not something we ever want
to do, or will do. I may rip out the code that does this soon.